requirements.txt
zyndai-agent
python-dotenv
gunicorn
supabase
requests